requirements.txt
psycopg2-binary>=2.9
pydantic>=2.0
streamlit>=1.30
billboard.py>=7.0
altair>=5.0
